refactor(ReactMediaObject): separate components
Issue:

When I add tests, the test file will be gigantic.

Outcome/side effects:

Turns out I didn't add props.style to the Bd component in my last PR. I've done it now.
